Changeset added
Added a changeset that illustrates various ideas. First is the notion
of non-automatically rollbackable tags. It should be noted that
a rollback should always be provided.
Secondly the notion of a context. Context can be used to separate the
various purposes of changing data.

Thirdly, but this was is implicit, is the oppertunity to talk about
outputting sql instead of executing it directly on the database.